+++ src/DALHAL/Devices/GeneralOutputs/SinglePulseOutput/DALHAL_SinglePulseOutput_JSON_Schema.cpp
@@ -36,7 +36,7 @@ namespace DALHAL {
 
     namespace JsonSchema {
 
-        constexpr FieldHardwarePin pinField = { DALHAL_COMMON_CFG_NAME_PIN, FieldPolicy::Required, static_cast<uint8_t>(GPIO_manager::PinFunc::IN) };
+        constexpr FieldHardwarePin pinField = { DALHAL_COMMON_CFG_NAME_PIN, FieldPolicy::Required, (GPIO_manager::PinFunc::IN) };
         constexpr FieldUInt pulseLengthField = { DALHAL_KEYNAME_SINGLE_PULSE_OUTPUT_DEFAULT_PULSE_LENGHT, FieldPolicy::Optional, 1, 0, 500};
         
         constexpr FieldString activeLevelField = { "activeLevel", FieldPolicy::Optional, DALHAL_COMMON_CFG_VALUE_PIN_LEVEL_HIGH, activeLevelStrings, FieldString::AllowedValuesPolicy::IgnoreCase};
